Remove `ReferringSitesDataTemplateSelector`
